feat(dal): graph goes vroom
Profiling of slow rebases with lots of updates revealed that most of the
time was being spent in replace_references, both calculating the merkle
tree hash and cloning node weights in copy_node_index_by_id. This "copy
on write" logic no longer makes sense to me with the operational
transforms. And in any case, we are still copy on write (from the read
only graph to the working copy). This PR removes the "copy on write"
logic which no longer makes sense. Instead we mutate the graph in place,
and keep a list of touched nodes. Using that list of touched nodes, we
recalculate the merkle tree hash for the graph once, in
`cleanup_and_merkle_tree_hash`. The result is a massive speed up in
graph operations. Even huge rebases come in under a second, and usually
under 300ms. Small rebase operations stay very fast, typically under
50ms.