feat(naxum): add `RequiredReply` extractor
This change adds a new extractor type `RequiredReply` that can be used
by handler functions. It will ensure that a reply subject is present in
the message and if not a `NoReplyRejection` is returned.

The most likely use case for this extractor is when processing core NATS
messages where a reply inbox is expected. Note that there is also a
`Reply` extractor already present in naxum, however it returns an
`Option<Subject>`.

Example
-------

```rust
pub async fn process_message(
    RequiredReply(reply_subject): RequiredReply,
    Json(request): Json<MyRequest>,
) {
    // ...
}
````

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>