Further destroy the existence of "nilId"
This commit furthers the destruction of "nilId" throughout the codebase.
It does this by migrating as many dal-based identifiers from the "pk!"
macro to the "id!" macro as possible. The macros are extremely similar,
but the "id!" macro does not include the ability to create "nilIds"
easily.

All remaining locations using the old "pk!" macro have been marked with
a "TODO" comment indicating what's to come.

As a reminder, this macro even works for primary keys, just as the old
macro worked for identifiers in the past.

In addition to these changes both the "node.rs" and "job_failure.rs"
modules have been removed due to lack of use. The job failures table
will still be migrated in order to not interrupt existing SI
deployments, but it is effectively dead.

You may also notice that "AttributePrototypeSource" had its "EnumIter"
and "EnumString" derived traits removed. This is because they rely on
the identifiers to implement or derive "Default", which the new macro
does not provide an implementation for (intentionally).

Signed-off-by: Nick Gerace <nick@systeminit.com>