feat(pinga,rebaser,veritech): impl work-draining graceful shutdown
This change updates the naxum setup for Pinga/Rebaser/Veritech to use a
work-limiting concurrency and will allow work-in-flight to complete when
a graceful shutdown is requested (via a `SIGINT` or `SIGTERM`). The
shutdown behavior is now closer to that of the example provided in #4444
which delivered the updated naxum functionality.

Closes: ENG-2655

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>