Allow ordered task groups for graceful shutdown
This commit allows for ordered task groups, which are comprised of
shutdown token and a corresponding tracker, when performing graceful
shutdown. In other words, when performing graceful shutdown, we will
now fully drain task groups in order. If using graceful shutdown
functionality from the "si-service" crate, then the timeout condition
will apply to all task groups.

Future considerations:

In the future, we may want to create a manager struct to manage task
groups as well as handle the shutdown for them. We may also want to
switch from a flat dependency list to a dependency tree in case we want
parallelism during shutdown.

Using a manager struct could also help with naming task groups. It could
be helpful in the future to know which task group is being drained (by
name) as well as what order that happens in.

Additional changes:

This commit also deletes some dead comments as well as increases some
shutdown logging to "info" level. In addition to that, we no longer
clone the tracker used for creating LayerDB clients; we borrow it.

Signed-off-by: Nick Gerace <nick@systeminit.com>