Annihilate reconciliation func remnants
This commit removes remnants of reconciliation funcs in the codebase.
Since they are unused and since their re-introduction would likely take
a different shape, we are destroying what remains. We loved you!

This is a second attempt (#4523 and #4454) and now works with existing
graphs due to deserializing enums that contained reconciliation funcs.

Signed-off-by: Nick Gerace <nick@systeminit.com>