fix(si-layer-cache): deserialize on blocking pool
When receiving writes from other services we have to deserialize them to
load them into the memory cache. But deserialization of large values is
(relatively) slow, and can lock up the tokio reactor, causing unexpected
issues elsewhere. This puts the deserialize ops on the blocking pool so
that other tasks can continue. The "slow runtime" is not available in
the layer cache. We may want to factor that out into its own crate so
that non-Dal crates can use it.