New SiNodeWeight traits & blanket impls
For a while now, we've known that there is at least one trait that would
be appropriate to extract from the current node weight implementations;
we just haven't known exactly what shape the trait(s) would have.

This is probably still not what the fully formed traits will look like,
but it is a starting point, and allows for some default blanket
implementations of other traits that we have already created around node
weights. It also allowed for a fair amount of code deduplication &
cleanup around the upcoming new node weights.