feat: API test structural changes
Various changes to the libraries + call pattern in the application. Adds a structured JSON output based on the execution of the tests itself. Allows the users to also specify individual tests which is helpful if you're developing a new one you don't have to run the whole collection all the time.