Remove metadata field in billing event
This commit removes the metadata field in billing events due to its
potential unbounded size. While the metadata retrieval logic is still in
play (mainly due to the "count"), it is unused. In the future, it should
be batched up and sent in a another process or thread to adhere to
data warehouse design conventions.

In addition to these changes, the ID of the func run is now recorded
for both resource creations and resource deletions.

Signed-off-by: Nick Gerace <nick@systeminit.com>