Will revert: sleep after going into maintenance
This commit sleeps for 120 after going into maintenance mode for sdf
because we have a weird deployment scenario for the previous commit
(7a1fc88). We need to make sure we properly drain all running tasks
and then delete all dead NATS Jetstream streams (done outside of the
code). After deploying and making the edits we need, we should revert
this commit.

Co-authored-by: John Watson <john@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>