fix: ensure only one is_default: true edge for any source
This will ensure we remove any is_default: true edges and convert them
to is_default: false if a NewEdge update for is_default: true comes in,
for any source. This is a little more highly coupled to the updates than
the transform correction but solves the problem for any source type. I
could not produce this issue locally in a test, so i'm not sure if it's
even possible, but this does not break schema variants during authoring.