feat: on demand asset installation
If the `on_demand_assets` feature flag is enabled for a user, all new
workspaces created by that user will have just intrinsics installed, and
assets will be installed at the moment that the component for that asset
is dropped on the canvas.

The local asset cache for sdf has to be updated manually by going to the
Admin dashboard.