Tune down sdf migrations logging
This commit tunes down sdf migrations logging in three areas. In all
areas, the logging is tuned to debug-level.

The first change involves logging when migrating individual modules.
At the time of writing, these modules take ~0.00-0.01 seconds to migrate
each. In addition, there are now over 100 modules installed, so the
logging spammed is not only no longer useful, but is also particularly
noisy.

The second change involves logging when modules have already been
installed. Currently, this is a warning-level log and now, the message
has been re-written to indicate that we are explicitly skipping the
installation.

The third change involves logging when importing a schema from a module.
This area is potentially the area most sensitive to change as it will
also change the log level for installed modules. This is a low risk
change as info-level logging should be performed at a higher level API
rather than "import_change_set" as it is used in many locations and for
multiple purposes. As part of this change, I looked through the usage
locations and existing tracing methods should be satisfactory.

What about performance regressions? We will still info-level log the
total elapsed time, so we can track the "elapsed" field in the trace if
need be. Yes, that potentially gives us a blind spot to regressions
based on total module count (e.g. introducing a performance regression
where the logarithmic curve in migrating each module becomes
progressively slower), but we should not be relying on these logs for
that kind of analysis anyway.

Additional change: this commit deletes some commented out and dead code.
In addition, it makes the "tick" logging more clear when migrating
builtins in sdf.

Signed-off-by: Nick Gerace <nick@systeminit.com>