Polish selection details for asset details panel
This commit polishes selection details and options for the asset details
panel. It uses the input socket icon pointing left as the icon for the
identity function.

These changes are powered by extending VormInput to use optgroup under
the hood to do so. When attaching funcs, SelectMenu mimics the new
optgroup behavior too.

In addition, the "color" option in the asset details panel now is
captialized to "Color" and the select box outline is much more subtle
when attaching funcs.

Co-authored-by: John Obelenus <jobelenus@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>