fix: don't write all dvus as finished by default
A bug in the implementation of the component concurrency limit meant we
were writing all dvu roots as "finished", but luckily still executing
them (i'm not really sure how it worked!). But this broke some transform
corrections related to DVUs, which was noticed when fixing the component
deletion bug. This ensures we write the roots correctly and restores the
component concurrency limit test.