Upsert inferred edges on creation and detach
This commit informs the frontend of new, inferred edges on component
creation. This happens when creating a component and directly placing
it into a frame.

This commit also handles an edge case for mutated, inferred connnections
when detaching children of frames.

Example scenario: there's a grandparent up-frame, a parent down-frame
and a child component or up-frame. If the parent frame has an input
socket with an arity of one and both the child and grandparent have an
output socket with a matching annotation, then the edge will be mutated.
The parent will now have a different value at its input socket because
the source side of the edge has changed.

Co-authored-by: Brit Myers <brit@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>

Inform frontend of mutated edges on detach

Co-authored-by: Brit Myers <brit@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>