feat: add management function backend and refactor veritech
Adds a new function backend to veritech/lang-js and connnects it to the
dal. No management functions exist yet so this code path will not yet be
excercised.

Refactors cyclone/veritech's request path a bit to make it simpler to
add new function types. Now just add a new variant to the
VeritechRequest enum and implement the CycloneRequestable trait for your
request payload type. Then let the compiler guide you to the places
where that enum is checked. That will get you all the way up to the
final stage, adding the new function code to lang-js.