Mock audit logs for frontend
This commit adds mocked audit logs for the frontend. Currently they are
based in the dal, though it is likely that they will move to their
own crate.

It's not in this commit, but there is a difference between a
"frontend type" audit log and the audit log that will be ultimately
be stored in the database. The purpose of this commit is to unlock
frontend creativity by providing generated audit logs. It is not meant
to reflect the structure that will ultimately be used by backend
services.

Co-authored-by: Wendy Bujalski <wendy@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>