chore: add a test to prevent breaking graph deserializations
Commits a small snapshot to the repository that has a graph with one
instance of every node weight and one of every edge weight kind, then
ensures we can deserialize it. This is a first pass at ensuring code
cannot be merged that will break graph deserialization.

Another test is necessary that writes one of every content object
somewhere and attempts to deserialize all of them.

This test graph can be generated at any time by running

  buck2 run //lib/dal:test-integration -- write_deserialization_data --ignored

Then you just have to update the constant in

lib/dal/tests/integration_test/deserialize/mod.rs

to point to the new graph.