chore: refactor the way we send a force change set response
Instead of constructing an axum Response by hand and adding the
force_change_set_id header manually, this introduces a new response type
for our sdf routes that handles the header automatically in its
implementation of the IntoResponse trait: ForceChangeSetResponse. Use
it anywhere you need to notify the frontend that the change set has
changed.