Replace "track resource counts" lambda with a view.
This is slower to run (10-11 seconds vs 3), but reports fresher data
and has fewer moving parts to fail. Performance can be improved later as
needed with indexes and materialized views.

Also exposes per-workspace resource counts for each
resource-hour, for diagnostic purposes and so that
we can (eventually) segment invoices by workspace.