Add filtering and pagination to fake audit logs
This commit adds filtering and pagination to fake audit logs that were
first introduced in commit 51f424a.

It performs a "second pass" filtering and pagination run on a generated
list of audit logs. Not only that, but that commit also polishes audit
log generation log, standardizes users and makes random number
generation easier to follow.

Signed-off-by: Nick Gerace <nick@systeminit.com>