Start skeleton of WorkspaceSnapshot & WorkspaceSnapshotGraph traits for domain object extensions
The idea is that that we should move as much of logic that is directly
interacting with `NodeWeight` & edges/`EdgeWeight` as close to the
versioned `WorkspaceSnapshotGraph` as possible, restricting things
"outside" of `WorkspaceSnapshot` to only using interfaces exposed throug
`Ext` traits on `WorkspaceSnapshot` that do not allow direct
manipulation of the graph.

By limiting things that are directly interacting with the graph
structure to things on `WorkspaceSnapshot` (or ideally
`WorkspaceSnapshotGraph`), we can hopefully limit the amount of
duplicated (and diverging) implementations of various logic.