Orphan all children before removing old frame version on schema version upgrade
While attempting to (re)connect the children of a frame to the newly
upgraded version, we try to find the `OutputSocket` `AttributeValue`s
from the old version, but this has already been removed. Rather than
creating specialized (re)connect logic for the schema version upgrade
scenario, we can have the existing logic treat it as a brand new
parent/child relationship by orphaning the children before we remove the
old version of the parent frame.