Consume and publish pending audit logs
This commit consumes and publishes pending audit logs. It contains a
basic audit logging flow for three scenarios: component creation,
property editor value updates, and component deletion. Through this
work, SI will now use the pending events architecture to flush pending
events when the rebaser "blesses" a set of events (or when there's
nothing to rebase).

Primary changes:

There's a new stream in town and it's the audit logs stream. This will
be where audit logs live after being consumed from the pending events
stream and published accordingly.

How do we know when audit logs are ready to be shuttled to the new
stream? There's two ways. The first is the most common path: if a rebase
is request, the rebaser-client will publish a final message on the audit
log subject for the given event session on the pending events stream.
After that, the rebaser-server will consume on the pending events stream
for a given event session, see all messages until the final message,
and them shuttle them to the audit logs stream. As for the second way,
if there's nothing to rebase, the DalContext will shuttle its own
pending events upon commit. In either case, the shuttle instance needs a
"final message" (denoted by aheader key).

Secondary changes:

As part of this architectural shift, we need to increment the rebase
request version. It now contains the EventSessionId.

We will also use a beta version of both the pending events stream and
the audit logs stream in order to ensure the behavior is what we expect
in a live, long-running environment.

Additional changes:

- Add ability for DalContext to make JetStream contexts
- Make all AuditLogKinds present tense
- Re-order all imports in "Cargo.toml" for all relevant crates in this
  commit
- Rename "shuttle" to "shuttle-server" and spin out relevant constant
  to a new crate: "shuttle-core"
- Version AuditLogs and AuditLogKinds starting with V1

Caveats:

Audit logging is skipped at the deprecated, "universal" tenancy level.

Context:

This work is the successor to #4814. Relevant PRs to help enable this
work include the following: #4852, #4854, #4872.

Signed-off-by: Nick Gerace <nick@systeminit.com>