Add shuttle crate
This commit provides "shuttle", a new crate and service that "shuttles"
or "forwards" messages from a limits-based NATS JetStream stream to a
given NATS subject.

This is needed for the pending events architecture and audit logging
work in order for the rebaser to drain pending events and publish them
to appropriate subjects.

Relevant PRs: #4814, #4852, #4854

Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Signed-off-by: Nick Gerace <nick@systeminit.com>