Replace mocked audit log data with real data
This commit replaces mocked audit log data with real data. It moves
AuditLog to "V2", but all active streams are "beta" streams, so
backwards compatibility should only matter for the lifespan of the
future flag and the "beta" streams.

This commit also reduces the events down to just creating a component,
updating a property editor value and deleting a component. Other event
kinds have been removed and AuditLog versions now exist in their own
submodules.

In addition to these changes, the frontend has adapted to the new audit
log shape. However, the filters and dropdowns are broken, which we will
fix as we determine what data should look like and how to interact with
it.

The audit log page now has a loader for reloading the set of audit logs
and contains a few changes to the headers. Unused fields have been
dropped for now too.

Signed-off-by: Nick Gerace <nick@systeminit.com>