Solidify audit log structure in the UI
This commit solidifies the audit log structure in the UI. In order to
do this, there is now an AuditLogV3 (reminder: backwards compatibility
only matters for the "beta" streams and we can reset the counter once
we switch over to the production streams).

Now, we have a display name for the event, an entity type for the event
and an entity name for the event.

This commit re-orders the columns and headers as well as adds selected
and hover background coloration to the rows. It does not contain fixes
for the filters nor are the selected and hover colors final. In those
verticals, this commit is just another step towards where we want to be.

Signed-off-by: Nick Gerace <nick@systeminit.com>