Allow billing to support more than 2 subscriptions.
The SQL changes here primarily allow us to support more than two subscriptions, so
that you can switch between multiple plans. Right now we support only free
trial and pay-as-you-go plans. This allows us to support custom Enterprise plans and
to move SI internal users to a new plan that doesn't generate invoices.

This also adds lambda changes that allow the lambda functions to be run more easily locally
(by setting up a venv).

This also fixes a bug where we would report the last day's usage of the free trial against
your pay-as-you-go plan. It would not have been invoiced, but it's inaccurate and would have
caused larger problems as we move between two plans that *do* charge. The issue was that the
end date of the free trial is a *date* and therefore 1 day off from the start date of the
pay-as-you-go subscription. We want to report all the hours during that gap day to the free trial.