Fix and expand the audit log headers
This commit fixes the audit log headers and expands their functionality.
Now, all headers are populated from live or compile-time data (other
than user invitations) to ensure that the application has up-to-date
headers.

This commit also introduces a new filter for the entity type. Now, you
can filter by entity type and/or the raw event kind. The raw event kind
is viewable by hovering over the title in the "Event" section. Change
Set and User filtering now works as intended too with a streamlined
filtering scheme for the System actor (thank you serde untagged!).

Other changes include matching the expand button's style to that of
the ellipsis on the action card and removing old "ip" header references
in the audit log UI code. The max message count in the audit logging
list call has been increased from 1000 to 10000 as well, though this
ultimately needs to be replaced with real pagination soon.

Timestamp refreshing:

This commit introduces timestamp refreshing and its first user is the
audit logs UI. Every sixty seconds, the timestamp string will refresh.
What was "17 minutes ago" will become "18 minutes ago", as an example.
In the future, we may want to tie this to a clock for more precision.

Audit logging integration test fix:

The audit logging integration test now accounts for a busy CI
environment and wraps the "source stream has no messages" check into a
new test helper function with a timeout and interval pattern.

Known issue:

With smaller windows, the leftmost filter dropdown can clip off the
screen. This is due to anchoring the DropdownMenu to the "th" element.
This requires a further look, but should not hinder users of medium or
larger windows.

Signed-off-by: Nick Gerace <nick@systeminit.com>