Move filtering and pagination into audit log list
This commit moves the filtering and pagination logic into the audit log
list function. The fake pagination logic is still in place, but this
moves us a step closer towards real pagination by ensuring all callers
go through the same interface.

This commit also makes the message fetch dynamic based on the number of
messages on the screen. Now, users and change sets are cached while
filtering audit logs.

The hold-up for proper pagination? Reverse timestamp sort. Consuming
off the stream provides a linearized view of audit logs. This is great
when you know where you want to start and end on a stream, but we do
not have that luxury here since we need to know how many logs fit our
filtered criteria as well as other relevant information.

Signed-off-by: Nick Gerace <nick@systeminit.com>