feat: mgmt fns can set component type, and set parent for components
Adds an optional `parent` property to creates and updates returned by
management functions that sets the frame parent. Also allows you to
change a component into a frame by setting the `root/si/type` property
of a component.

Component geometries returned and set via management functions are now
all relative to the "manager" component, which gets x and y of 0.0. To
place a component to the left or above the manager, use negative
geometry. To the right or below, positive geometry. This allows reliably
placing components inside the parent frame in any view.